parse smallint as shortype

+++ src/main/scala/com/databricks/spark/redshift/RedshiftJDBCWrapper.scala
@@ -332,7 +332,7 @@ private[redshift] class JDBCWrapper {
       case java.sql.Types.REAL          => DoubleType
       case java.sql.Types.REF           => StringType
       case java.sql.Types.ROWID         => LongType
-      case java.sql.Types.SMALLINT      => IntegerType
+      case java.sql.Types.SMALLINT      => ShortType
       case java.sql.Types.SQLXML        => StringType
       case java.sql.Types.STRUCT        => StringType
       case java.sql.Types.TIME          => TimestampType
